Added sysroot support to ease cross compilation
Users can specify sysroot during the configuration step.
The system will try to detect the target and architecture of the sysroot and pass on the information to all Clang compilers.
The detected toolchains allow to generate binaries for alternate architectures without fully installing the OS.

Tested with FreeBSD and Debian sysroots, arm/aarch64/ppc/amd64