OpenCL parsing and code generation progress
OpenCL typing system allows for generic address space. When methods are called, the code generation should replace the generic address space with the concrete one.
It requires all types to be resolved first, then all methods to be "instantiated".
With this commit, instructions are now resolved in the scope of the method and are prepared for instantiation.