GLRP: hooked up table generation
Table generated by the LALR algorithm is pickled, gzipped and converted to
base64 so it remains in text format. Parser creation can either test the
pickled grammar or force the generation of the table file.
Table files are not yet added in git as they can be generated on demand,
and are expected to change often in the next commits.