Fixed counterexample calculation
Counterexample calculation could fail when discovering the lookahead in a state, but the common parent could be found only in a parent state.
Discovering the lookahead in a state would add all discovered node in the "seen" set, preventing a new recursion to discover the parent states.
The new algorithm discovers all paths (the current state and the parents) and will queue the parents to avoid calling the backtrack function a second time.