Added more efficient implementation of Semaphores
Semaphores on pthread does not allow Post to release more than one thread,
which causes the application performance to depend on the number of
threads. Implementation was changed to depend on a lower level primitive
(futex) on Linux to be able to release several workers at once.
This uncovered an implementation issue for 32-bits interlocked on ARM64,
which is also fixed.