Ensure proper acquire/release semantics on ARM
Fixed a crash on ARM due to invalid implementation of acquire/release semantics.
Interlocked operation semantics should follow the semantics detailed in the
thread sanitizer implementation, as those semantics are tested and reported to
be valid by thread sanitizer.
ARM was missing a release barrier on set_conditional which lead to crashes in
the load test of the task scheduler.