pyxx: Refactored decl-specifier-seq
In order to avoid an ambiguity between a declarator and a type name, the
decl-specifier-seq rule is broken down in two. The first half may include a
type name and ends when encountering a defining type specifier.
The second half may not include a type name as this would be interpreted
as a declarator instead.