Add pragma once in certain header files to help IDEs
Some header files that include ther implementation in another header file cause issues in certain IDEs.
The only solution that seems to work on QtCreator, VsCode and CLion is to use header guards in the header file, include the header file in the inline file and use header guards AND pragma once in the inline file.