changed refptr and scopedptr to not be intrusive pointers.
refptr and scopedptr retain almost the same API but refptr does not necessitate the class to inherit from refcountable anymore. Classes do not need to remember their allcoator either.
Both refptr and scopedptr are now the size of two pointers (one for the control block and one for the object pointed to)
weakptr remains an intrusive pointer as an optimisation.

Refactored the filesystem and resource APIs to avoid addref/decref:
- Resources are stored as scoped pointers instead
- Filesystem API is changed so that the IO implementation is moved into the ticket (which is already refcounted) instead of the file. The IO is still run async and is independent of the lifetime of the file/folder objects. Files are stored as scoped pointers in the folder.